Refactor Bundle integration tests
