+++ test/smoke/suite_test.go
@@ -27,7 +27,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	trustapi "github.com/cert-manager/trust-manager/pkg/apis/trust/v1alpha1"
-	"github.com/cert-manager/trust-manager/pkg/bundle/controller"
 	"github.com/cert-manager/trust-manager/test"
 	"github.com/cert-manager/trust-manager/test/dummy"
 	"github.com/cert-manager/trust-manager/test/env"
@@ -65,9 +64,7 @@ var _ = Describe("Smoke", func() {
 		By("Creating Bundle for test")
 		testData := env.DefaultTrustData()
 
-		testBundle := env.NewTestBundleConfigMapTarget(ctx, cl, controller.Options{
-			Namespace: cnf.TrustNamespace,
-		}, testData)
+		testBundle := env.NewTestBundleConfigMapTarget(ctx, cl, cnf.TrustNamespace, testData)
 
 		By("Ensuring the Bundle has Synced")
 		env.EventuallyBundleHasSyncedAllNamespaces(ctx, cl, testBundle.Name, dummy.DefaultJoinedCerts())
@@ -84,9 +81,7 @@ var _ = Describe("Smoke", func() {
 		By("Creating Bundle for test")
 		testData := env.DefaultTrustData()
 
-		testBundle := env.NewTestBundleSecretTarget(ctx, cl, controller.Options{
-			Namespace: cnf.TrustNamespace,
-		}, testData)
+		testBundle := env.NewTestBundleSecretTarget(ctx, cl, cnf.TrustNamespace, testData)
 
 		By("Ensuring the Bundle has Synced")
 		env.EventuallyBundleHasSyncedAllNamespaces(ctx, cl, testBundle.Name, dummy.DefaultJoinedCerts())
